app/requirements.txt
azure-functions>=1.18.0
azure-identity>=1.20.0
azure-ai-projects>=1.0.0b11
azure-core>=1.30.0